pyproject.toml
[project]
name = "projeto-final"
version = "0.1.0"
description = ""
authors = [
    {name = "Felipe Alves",email = "lipealvessax@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "duckdb (>=1.3.2,<2.0.0)",
    "minio (>=7.2.16,<8.0.0)",
    "pyspark (==3.4.0)",
    "delta-spark (==2.4.0)",
    "boto3 (>=1.40.30,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
